Fix a number of instances of flickering (#86)
The most common cause of flickering is that there is a repaint
between sequences that are meant to be interpreted together.
This can be avoided by buffering sequences first and then
writing them out all at once. Do this, and add testing infrastructure
to verify that it stays this way.