Run OhMyREPL keybindings in a fixed world age
This should avoid most invalidations caused by loading other packages.

Also remove the NEW_KEYBINDINGS global as it's hard to follow where this
is added to.

Rearrange all overrides so the implementation code can be precompiled,
with the only thing which is `eval`d at runtime being the method
pirating itself.

Include global mutable "hook" binding for `LineEdit.refresh_line` to
allow us to fix the world age of the implementation.